+++ developer/src/common/web/utils/index.ts
@@ -1,5 +1,4 @@
 export { validateMITLicense } from './src/validate-mit-license.js';
 export { KeymanSentry } from './src/KeymanSentry.js';
 export { getOption, loadOptions, clearOptions } from './src/options.js';
-export * as Osk from './src/osk.js';
 export { escapeMarkdownChar } from './src/markdown.js';
