+++ developer/src/common/web/utils/tsconfig.json
@@ -5,15 +5,9 @@
       "rootDir": ".",
       "outDir": "./build/",
       "baseUrl": ".",
-      "paths": {
-        "@keymanapp/common-types": ["../../../../../common/web/types/src/main"],
-      },
     },
   "include": [
     "index.ts",
     "src/**/*.ts",
   ],
-  "references": [
-    { "path": "../../../../../common/web/types/" },
-  ]
 }
