+++ developer/src/kmc-analyze/src/osk-character-use/index.ts
@@ -1,6 +1,6 @@
 import { CompilerCallbacks, KeymanFileTypes, KvksFile, KvksFileReader, TouchLayout, TouchLayoutFileReader } from "@keymanapp/common-types";
-import { CompilerMessages } from '@keymanapp/kmc-kmn';
-import { escapeMarkdownChar, Osk } from '@keymanapp/developer-utils';
+import { CompilerMessages, Osk } from '@keymanapp/kmc-kmn';
+import { escapeMarkdownChar } from '@keymanapp/developer-utils';
 import { getOskFromKmnFile } from "../util/get-osk-from-kmn-file.js";
 import { AnalyzerMessages } from "../messages.js";
